wire up delete
